Upload CloudFormation YAML files
Add the CloudFormation YAML files and Dockerfile files (in code) to the
repository. These files are ultimately deployed to an S3 bucket
'access-idp-templates' which is used by CloudFormation to create a new
access-idp deployment, but keeping the files in the Git repo makes it
easier to track changes.

Also, update the README file for IdP v5.x.